Split WalletLink websocket filter helpers
